refactor so that we use entry scripts for setting up test environment
Looks like docker execution on linux (which is what Travis environment is) behaves a bit differently with regards to user permissions.  These changes are an attempt to make sure the container user is lined up to use the same uid/gid as what the host user is running the script.  Hopefully that will fix any user permission issues with running tests in travis.